fix(decompose): collapse splinter modules back into deep modules to clear subgraph-gate
Reverses recent file-decompositions that pushed touched communities above
their boundary-width / implicit-globals baselines:

  - measures/_subgraph_gate boundary-width  112 → ≤92
  - webapp/shell             boundary-width 1056 → ≤1037
  - webapp/shell             implicit-globals 862 → ≤853
  - graph-model/pure         boundary-width  464 → ≤461

Strategy per fp-rearchitecting/SKILL.md: tighten visibility instead of
refreshing baselines. Internal helpers become module-private functions
inside the deep module; test-only support moves under __tests__/ so
listProductionSources excludes it.

Files:
  - ast-purity-ratio.ts: 8 splinter files folded into one measure.ts
  - VoiceTreeGraphView.ts / view.ts: subfolder helpers inlined
  - voicetree-transcribe.tsx: types moved to presentation.tsx
  - delete-and-merge-filesystem.test, addProgressNodeMcp.test:
    test-support moved into __tests__/
  - parse-markdown-to-node.test: fixtures inlined via readFileSync
